requirements.txt
yt-dlp
requests
python-vlc
flask
flask-socketio
keyboard
browser-cookie3
certifi
urllib3
pycaw 
comtypes
python-mpv
redis